+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestOzoneManagerHAWithData.java
@@ -17,28 +17,15 @@
 package org.apache.hadoop.ozone.om;
 
 import java.io.IOException;
-import java.lang.management.ManagementFactory;
 import java.net.ConnectException;
-import java.net.InetSocketAddress;
-import java.time.Instant;
 import java.util.ArrayList;
-import java.util.BitSet;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.TreeSet;
 import java.util.UUID;
 
-import org.apache.hadoop.ozone.OzoneAcl;
 import org.apache.hadoop.ozone.OzoneConfigKeys;
 import org.apache.hadoop.ozone.client.OzoneMultipartUploadPartListParts;
-import org.apache.hadoop.ozone.om.ratis.OzoneManagerRatisServer;
-import org.apache.hadoop.ozone.security.acl.OzoneObj;
-import org.apache.hadoop.ozone.security.acl.OzoneObjInfo;
-import org.apache.hadoop.security.UserGroupInformation;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -47,28 +34,21 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.rules.Timeout;
-import org.apache.log4j.Logger;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.hadoop.hdds.client.ReplicationFactor;
 import org.apache.hadoop.hdds.client.ReplicationType;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-import org.apache.hadoop.hdds.protocol.StorageType;
-import org.apache.hadoop.hdfs.LogVerificationAppender;
 import org.apache.hadoop.ipc.RemoteException;
 import org.apache.hadoop.ozone.MiniOzoneCluster;
 import org.apache.hadoop.ozone.MiniOzoneHAClusterImpl;
-import org.apache.hadoop.ozone.OzoneTestUtils;
-import org.apache.hadoop.ozone.client.BucketArgs;
 import org.apache.hadoop.ozone.client.ObjectStore;
 import org.apache.hadoop.ozone.client.OzoneBucket;
-import org.apache.hadoop.ozone.client.OzoneClient;
 import org.apache.hadoop.ozone.client.OzoneKeyDetails;
 import org.apache.hadoop.ozone.client.io.OzoneInputStream;
 import org.apache.hadoop.ozone.client.io.OzoneOutputStream;
 import org.apache.hadoop.ozone.om.exceptions.OMException;
 import org.apache.hadoop.ozone.om.ha.OMFailoverProxyProvider;
-import org.apache.hadoop.ozone.om.ha.OMProxyInfo;
 import org.apache.hadoop.ozone.om.helpers.OmMultipartInfo;
 import org.apache.hadoop.ozone.om.helpers.OmMultipartUploadCompleteInfo;
 import org.apache.hadoop.test.GenericTestUtils;
@@ -83,33 +63,21 @@ import static org.apache.hadoop.fs.CommonConfigurationKeysPublic
 
 import static org.apache.hadoop.ozone.MiniOzoneHAClusterImpl
     .NODE_FAILURE_TIMEOUT;
-import static org.apache.hadoop.ozone.OzoneAcl.AclScope.ACCESS;
-import static org.apache.hadoop.ozone.OzoneAcl.AclScope.DEFAULT;
 import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_ACL_ENABLED;
 import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_ADMINISTRATORS_WILDCARD;
 import static org.apache.hadoop.ozone.OzoneConfigKeys
     .OZONE_CLIENT_FAILOVER_MAX_ATTEMPTS_KEY;
 import static org.apache.hadoop.ozone.OzoneConfigKeys
-    .OZONE_CLIENT_WAIT_BETWEEN_RETRIES_MILLIS_DEFAULT;
-import static org.apache.hadoop.ozone.OzoneConfigKeys
     .OZONE_OPEN_KEY_EXPIRE_THRESHOLD_SECONDS;
 import static org.apache.hadoop.ozone.om.exceptions.OMException.ResultCodes.DIRECTORY_NOT_FOUND;
 import static org.apache.hadoop.ozone.om.exceptions.OMException.ResultCodes.FILE_ALREADY_EXISTS;
 import static org.apache.hadoop.ozone.om.exceptions.OMException.ResultCodes.NOT_A_FILE;
-import static org.apache.hadoop.ozone.security.acl.IAccessAuthorizer.ACLIdentityType.USER;
-import static org.apache.hadoop.ozone.security.acl.IAccessAuthorizer.ACLType.READ;
-import static org.apache.hadoop.ozone.security.acl.IAccessAuthorizer.ACLType.WRITE;
-import static org.apache.ratis.server.metrics.RatisMetrics.RATIS_APPLICATION_NAME_METRICS;
 import static org.junit.Assert.fail;
-
-import javax.management.MBeanInfo;
-import javax.management.MBeanServer;
-import javax.management.ObjectName;
 
 /**
  * Test Ozone Manager operation in distributed handler scenario.
  */
-public class TestOzoneManagerHA {
+public class TestOzoneManagerHAWithData {
 
   private MiniOzoneHAClusterImpl cluster = null;
   private ObjectStore objectStore;
@@ -176,83 +144,6 @@ public class TestOzoneManagerHA {
     if (cluster != null) {
       cluster.shutdown();
     }
-  }
-
-
-  private OzoneVolume createAndCheckVolume(String volumeName)
-      throws Exception {
-    String userName = "user" + RandomStringUtils.randomNumeric(5);
-    String adminName = "admin" + RandomStringUtils.randomNumeric(5);
-    VolumeArgs createVolumeArgs = VolumeArgs.newBuilder()
-        .setOwner(userName)
-        .setAdmin(adminName)
-        .build();
-
-    objectStore.createVolume(volumeName, createVolumeArgs);
-
-    OzoneVolume retVolume = objectStore.getVolume(volumeName);
-
-    Assert.assertTrue(retVolume.getName().equals(volumeName));
-    Assert.assertTrue(retVolume.getOwner().equals(userName));
-    Assert.assertTrue(retVolume.getAdmin().equals(adminName));
-
-    return retVolume;
-  }
-  @Test
-  public void testAllVolumeOperations() throws Exception {
-
-    String volumeName = "volume" + RandomStringUtils.randomNumeric(5);
-
-    createAndCheckVolume(volumeName);
-
-    objectStore.deleteVolume(volumeName);
-
-    OzoneTestUtils.expectOmException(OMException.ResultCodes.VOLUME_NOT_FOUND,
-        () -> objectStore.getVolume(volumeName));
-
-    OzoneTestUtils.expectOmException(OMException.ResultCodes.VOLUME_NOT_FOUND,
-        () -> objectStore.deleteVolume(volumeName));
-  }
-
-
-  @Test
-  public void testAllBucketOperations() throws Exception {
-
-    String volumeName = "volume" + RandomStringUtils.randomNumeric(5);
-    String bucketName = "volume" + RandomStringUtils.randomNumeric(5);
-
-    OzoneVolume retVolume = createAndCheckVolume(volumeName);
-
-    BucketArgs bucketArgs =
-        BucketArgs.newBuilder().setStorageType(StorageType.DISK)
-            .setVersioning(true).build();
-
-
-    retVolume.createBucket(bucketName, bucketArgs);
-
-
-    OzoneBucket ozoneBucket = retVolume.getBucket(bucketName);
-
-    Assert.assertEquals(volumeName, ozoneBucket.getVolumeName());
-    Assert.assertEquals(bucketName, ozoneBucket.getName());
-    Assert.assertTrue(ozoneBucket.getVersioning());
-    Assert.assertEquals(StorageType.DISK, ozoneBucket.getStorageType());
-    Assert.assertFalse(ozoneBucket.getCreationTime().isAfter(Instant.now()));
-
-
-    // Change versioning to false
-    ozoneBucket.setVersioning(false);
-
-    ozoneBucket = retVolume.getBucket(bucketName);
-    Assert.assertFalse(ozoneBucket.getVersioning());
-
-    retVolume.deleteBucket(bucketName);
-
-    OzoneTestUtils.expectOmException(OMException.ResultCodes.BUCKET_NOT_FOUND,
-        () -> retVolume.deleteBucket(bucketName));
-
-
-
   }
 
   /**
@@ -672,449 +563,6 @@ public class TestOzoneManagerHA {
     }
   }
 
-  /**
-   * Test that OMFailoverProxyProvider creates an OM proxy for each OM in the
-   * cluster.
-   */
-  @Test
-  public void testOMProxyProviderInitialization() throws Exception {
-    OzoneClient rpcClient = cluster.getRpcClient();
-    OMFailoverProxyProvider omFailoverProxyProvider =
-        rpcClient.getObjectStore().getClientProxy().getOMProxyProvider();
-    List<OMProxyInfo> omProxies =
-        omFailoverProxyProvider.getOMProxyInfos();
-
-    Assert.assertEquals(numOfOMs, omProxies.size());
-
-    for (int i = 0; i < numOfOMs; i++) {
-      InetSocketAddress omRpcServerAddr =
-          cluster.getOzoneManager(i).getOmRpcServerAddr();
-      boolean omClientProxyExists = false;
-      for (OMProxyInfo omProxyInfo : omProxies) {
-        if (omProxyInfo.getAddress().equals(omRpcServerAddr)) {
-          omClientProxyExists = true;
-          break;
-        }
-      }
-      Assert.assertTrue("There is no OM Client Proxy corresponding to OM " +
-              "node" + cluster.getOzoneManager(i).getOMNodeId(),
-          omClientProxyExists);
-    }
-  }
-
-  /**
-   * Test OMFailoverProxyProvider failover on connection exception to OM client.
-   */
-  @Ignore("This test randomly failing. Let's enable once its fixed.")
-  @Test
-  public void testOMProxyProviderFailoverOnConnectionFailure()
-      throws Exception {
-    OMFailoverProxyProvider omFailoverProxyProvider =
-        objectStore.getClientProxy().getOMProxyProvider();
-    String firstProxyNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    createVolumeTest(true);
-
-    // On stopping the current OM Proxy, the next connection attempt should
-    // failover to a another OM proxy.
-    cluster.stopOzoneManager(firstProxyNodeId);
-    Thread.sleep(OZONE_CLIENT_WAIT_BETWEEN_RETRIES_MILLIS_DEFAULT * 4);
-
-    // Next request to the proxy provider should result in a failover
-    createVolumeTest(true);
-    Thread.sleep(OZONE_CLIENT_WAIT_BETWEEN_RETRIES_MILLIS_DEFAULT);
-
-    // Get the new OM Proxy NodeId
-    String newProxyNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    // Verify that a failover occured. the new proxy nodeId should be
-    // different from the old proxy nodeId.
-    Assert.assertNotEquals("Failover did not occur as expected",
-        firstProxyNodeId, newProxyNodeId);
-  }
-
-  /**
-   * Test OMFailoverProxyProvider failover when current OM proxy is not
-   * the current OM Leader.
-   */
-  @Test
-  public void testOMProxyProviderFailoverToCurrentLeader() throws Exception {
-    OMFailoverProxyProvider omFailoverProxyProvider =
-        objectStore.getClientProxy().getOMProxyProvider();
-
-    // Run couple of createVolume tests to discover the current Leader OM
-    createVolumeTest(true);
-    createVolumeTest(true);
-
-    // The OMFailoverProxyProvider will point to the current leader OM node.
-    String leaderOMNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    // Perform a manual failover of the proxy provider to move the
-    // currentProxyIndex to a node other than the leader OM.
-    omFailoverProxyProvider.performFailoverToNextProxy();
-
-    String newProxyNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-    Assert.assertNotEquals(leaderOMNodeId, newProxyNodeId);
-
-    // Once another request is sent to this new proxy node, the leader
-    // information must be returned via the response and a failover must
-    // happen to the leader proxy node.
-    createVolumeTest(true);
-    Thread.sleep(2000);
-
-    String newLeaderOMNodeId =
-        omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    // The old and new Leader OM NodeId must match since there was no new
-    // election in the Ratis ring.
-    Assert.assertEquals(leaderOMNodeId, newLeaderOMNodeId);
-  }
-
-  @Test
-  public void testOMRetryProxy() throws Exception {
-    // Stop all the OMs.
-    for (int i = 0; i < numOfOMs; i++) {
-      cluster.stopOzoneManager(i);
-    }
-
-    final LogVerificationAppender appender = new LogVerificationAppender();
-    final org.apache.log4j.Logger logger = Logger.getRootLogger();
-    logger.addAppender(appender);
-
-    try {
-      createVolumeTest(true);
-      // After making N (set maxRetries value) connection attempts to OMs,
-      // the RpcClient should give up.
-      fail("TestOMRetryProxy should fail when there are no OMs running");
-    } catch (ConnectException e) {
-      Assert.assertEquals(1,
-          appender.countLinesWithMessage("Failed to connect to OMs:"));
-      Assert.assertEquals(OZONE_CLIENT_FAILOVER_MAX_ATTEMPTS,
-          appender.countLinesWithMessage("Trying to failover"));
-      Assert.assertEquals(1,
-          appender.countLinesWithMessage("Attempted " +
-              OZONE_CLIENT_FAILOVER_MAX_ATTEMPTS + " failovers."));
-    }
-  }
-
-  @Test
-  public void testReadRequest() throws Exception {
-    String volumeName = "volume" + RandomStringUtils.randomNumeric(5);
-    objectStore.createVolume(volumeName);
-
-    OMFailoverProxyProvider omFailoverProxyProvider =
-        objectStore.getClientProxy().getOMProxyProvider();
-    String currentLeaderNodeId = omFailoverProxyProvider
-        .getCurrentProxyOMNodeId();
-
-    // A read request from any proxy should failover to the current leader OM
-    for (int i = 0; i < numOfOMs; i++) {
-      // Failover OMFailoverProxyProvider to OM at index i
-      OzoneManager ozoneManager = cluster.getOzoneManager(i);
-
-      // Get the ObjectStore and FailoverProxyProvider for OM at index i
-      final ObjectStore store = OzoneClientFactory.getRpcClient(
-          omServiceId, conf).getObjectStore();
-      final OMFailoverProxyProvider proxyProvider =
-          store.getClientProxy().getOMProxyProvider();
-
-      // Failover to the OM node that the objectStore points to
-      omFailoverProxyProvider.performFailoverIfRequired(
-          ozoneManager.getOMNodeId());
-
-      // A read request should result in the proxyProvider failing over to
-      // leader node.
-      OzoneVolume volume = store.getVolume(volumeName);
-      Assert.assertEquals(volumeName, volume.getName());
-
-      Assert.assertEquals(currentLeaderNodeId,
-          proxyProvider.getCurrentProxyOMNodeId());
-    }
-  }
-
-  @Test
-  public void testAddBucketAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    OzoneAcl defaultUserAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.BUCKET)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName()).build();
-
-    testAddAcl(remoteUserName, ozoneObj, defaultUserAcl);
-  }
-  @Test
-  public void testRemoveBucketAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    OzoneAcl defaultUserAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.BUCKET)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName()).build();
-
-    testRemoveAcl(remoteUserName, ozoneObj, defaultUserAcl);
-
-  }
-
-  @Test
-  public void testSetBucketAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    OzoneAcl defaultUserAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.BUCKET)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName()).build();
-
-    testSetAcl(remoteUserName, ozoneObj, defaultUserAcl);
-  }
-
-  private boolean containsAcl(OzoneAcl ozoneAcl, List<OzoneAcl> ozoneAcls) {
-    for (OzoneAcl acl : ozoneAcls) {
-      boolean result = compareAcls(ozoneAcl, acl);
-      if (result) {
-        // We found a match, return.
-        return result;
-      }
-    }
-    return false;
-  }
-
-  private boolean compareAcls(OzoneAcl givenAcl, OzoneAcl existingAcl) {
-    if (givenAcl.getType().equals(existingAcl.getType())
-        && givenAcl.getName().equals(existingAcl.getName())
-        && givenAcl.getAclScope().equals(existingAcl.getAclScope())) {
-      BitSet bitSet = (BitSet) givenAcl.getAclBitSet().clone();
-      bitSet.and(existingAcl.getAclBitSet());
-      if (bitSet.equals(existingAcl.getAclBitSet())) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  @Test
-  public void testAddKeyAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    OzoneAcl userAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    String key = createKey(ozoneBucket);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.KEY)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName())
-        .setKeyName(key).build();
-
-    testAddAcl(remoteUserName, ozoneObj, userAcl);
-  }
-
-  @Test
-  public void testRemoveKeyAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    OzoneAcl userAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    String key = createKey(ozoneBucket);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.KEY)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName())
-        .setKeyName(key).build();
-
-    testRemoveAcl(remoteUserName, ozoneObj, userAcl);
-
-  }
-
-  @Test
-  public void testSetKeyAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    OzoneAcl userAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    String key = createKey(ozoneBucket);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.KEY)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName())
-        .setKeyName(key).build();
-
-    testSetAcl(remoteUserName, ozoneObj, userAcl);
-
-  }
-
-  @Test
-  public void testAddPrefixAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    String prefixName = RandomStringUtils.randomAlphabetic(5) +"/";
-    OzoneAcl defaultUserAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.PREFIX)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName())
-        .setPrefixName(prefixName).build();
-
-    testAddAcl(remoteUserName, ozoneObj, defaultUserAcl);
-  }
-  @Test
-  public void testRemovePrefixAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    String prefixName = RandomStringUtils.randomAlphabetic(5) +"/";
-    OzoneAcl userAcl = new OzoneAcl(USER, remoteUserName,
-        READ, ACCESS);
-    OzoneAcl userAcl1 = new OzoneAcl(USER, "remote",
-        READ, ACCESS);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.PREFIX)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName())
-        .setPrefixName(prefixName).build();
-
-    boolean result = objectStore.addAcl(ozoneObj, userAcl);
-    Assert.assertTrue(result);
-
-    result = objectStore.addAcl(ozoneObj, userAcl1);
-    Assert.assertTrue(result);
-
-    result = objectStore.removeAcl(ozoneObj, userAcl);
-    Assert.assertTrue(result);
-
-    // try removing already removed acl.
-    result = objectStore.removeAcl(ozoneObj, userAcl);
-    Assert.assertFalse(result);
-
-    result = objectStore.removeAcl(ozoneObj, userAcl1);
-    Assert.assertTrue(result);
-
-  }
-
-  @Test
-  public void testSetPrefixAcl() throws Exception {
-    OzoneBucket ozoneBucket = setupBucket();
-    String remoteUserName = "remoteUser";
-    String prefixName = RandomStringUtils.randomAlphabetic(5) +"/";
-    OzoneAcl defaultUserAcl = new OzoneAcl(USER, remoteUserName,
-        READ, DEFAULT);
-
-    OzoneObj ozoneObj = OzoneObjInfo.Builder.newBuilder()
-        .setResType(OzoneObj.ResourceType.PREFIX)
-        .setStoreType(OzoneObj.StoreType.OZONE)
-        .setVolumeName(ozoneBucket.getVolumeName())
-        .setBucketName(ozoneBucket.getName())
-        .setPrefixName(prefixName).build();
-
-    testSetAcl(remoteUserName, ozoneObj, defaultUserAcl);
-  }
-
-
-  private void testSetAcl(String remoteUserName, OzoneObj ozoneObj,
-      OzoneAcl userAcl) throws Exception {
-    // As by default create will add some default acls in RpcClient.
-
-    if (!ozoneObj.getResourceType().name().equals(
-        OzoneObj.ResourceType.PREFIX.name())) {
-      List<OzoneAcl> acls = objectStore.getAcl(ozoneObj);
-
-      Assert.assertTrue(acls.size() > 0);
-    }
-
-    OzoneAcl modifiedUserAcl = new OzoneAcl(USER, remoteUserName,
-        WRITE, DEFAULT);
-
-    List<OzoneAcl> newAcls = Collections.singletonList(modifiedUserAcl);
-    boolean setAcl = objectStore.setAcl(ozoneObj, newAcls);
-    Assert.assertTrue(setAcl);
-
-    // Get acls and check whether they are reset or not.
-    List<OzoneAcl> getAcls = objectStore.getAcl(ozoneObj);
-
-    Assert.assertTrue(newAcls.size() == getAcls.size());
-    int i = 0;
-    for (OzoneAcl ozoneAcl : newAcls) {
-      Assert.assertTrue(compareAcls(getAcls.get(i++), ozoneAcl));
-    }
-
-  }
-
-  private void testAddAcl(String remoteUserName, OzoneObj ozoneObj,
-      OzoneAcl userAcl) throws Exception {
-    boolean addAcl = objectStore.addAcl(ozoneObj, userAcl);
-    Assert.assertTrue(addAcl);
-
-    List<OzoneAcl> acls = objectStore.getAcl(ozoneObj);
-
-    Assert.assertTrue(containsAcl(userAcl, acls));
-
-    // Add an already existing acl.
-    addAcl = objectStore.addAcl(ozoneObj, userAcl);
-    Assert.assertFalse(addAcl);
-
-    // Add an acl by changing acl type with same type, name and scope.
-    userAcl = new OzoneAcl(USER, remoteUserName,
-        WRITE, DEFAULT);
-    addAcl = objectStore.addAcl(ozoneObj, userAcl);
-    Assert.assertTrue(addAcl);
-  }
-
-  private void testRemoveAcl(String remoteUserName, OzoneObj ozoneObj,
-      OzoneAcl userAcl)
-      throws Exception{
-    // As by default create will add some default acls in RpcClient.
-    List<OzoneAcl> acls = objectStore.getAcl(ozoneObj);
-
-    Assert.assertTrue(acls.size() > 0);
-
-    // Remove an existing acl.
-    boolean removeAcl = objectStore.removeAcl(ozoneObj, acls.get(0));
-    Assert.assertTrue(removeAcl);
-
-    // Trying to remove an already removed acl.
-    removeAcl = objectStore.removeAcl(ozoneObj, acls.get(0));
-    Assert.assertFalse(removeAcl);
-
-    boolean addAcl = objectStore.addAcl(ozoneObj, userAcl);
-    Assert.assertTrue(addAcl);
-
-    // Just changed acl type here to write, rest all is same as defaultUserAcl.
-    OzoneAcl modifiedUserAcl = new OzoneAcl(USER, remoteUserName,
-        WRITE, DEFAULT);
-    addAcl = objectStore.addAcl(ozoneObj, modifiedUserAcl);
-    Assert.assertTrue(addAcl);
-
-    removeAcl = objectStore.removeAcl(ozoneObj, modifiedUserAcl);
-    Assert.assertTrue(removeAcl);
-
-    removeAcl = objectStore.removeAcl(ozoneObj, userAcl);
-    Assert.assertTrue(removeAcl);
-  }
-
   @Test
   public void testOMRatisSnapshot() throws Exception {
     String userName = "user" + RandomStringUtils.randomNumeric(5);
@@ -1321,65 +769,6 @@ public class TestOzoneManagerHA {
     validateListParts(ozoneBucket, keyName, uploadID, partsMap);
 
   }
-
-  @Ignore("This test randomly failing. Let's enable once its fixed.")
-  @Test
-  public void testListVolumes() throws Exception {
-    String userName = UserGroupInformation.getCurrentUser().getUserName();
-    String adminName = userName;
-
-    Set<String> expectedVolumes = new TreeSet<>();
-    for (int i=0; i < 100; i++) {
-      String volumeName = "vol" + i;
-      expectedVolumes.add(volumeName);
-      VolumeArgs createVolumeArgs = VolumeArgs.newBuilder()
-          .setOwner(userName)
-          .setAdmin(adminName)
-          .build();
-      objectStore.createVolume(volumeName, createVolumeArgs);
-    }
-
-    validateVolumesList(userName, expectedVolumes);
-
-    // Stop leader OM, and then validate list volumes for user.
-    stopLeaderOM();
-    Thread.sleep(NODE_FAILURE_TIMEOUT * 2);
-
-    validateVolumesList(userName, expectedVolumes);
-
-  }
-
-  @Test
-  public void testJMXMetrics() throws Exception {
-    // Verify any one ratis metric is exposed by JMX MBeanServer
-    OzoneManagerRatisServer ratisServer =
-        cluster.getOzoneManager(0).getOmRatisServer();
-    ObjectName oname = new ObjectName(RATIS_APPLICATION_NAME_METRICS, "name",
-        RATIS_APPLICATION_NAME_METRICS + ".log_worker." +
-            ratisServer.getRaftPeerId().toString() + ".flushCount");
-    MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
-    MBeanInfo mBeanInfo = mBeanServer.getMBeanInfo(oname);
-    Assert.assertNotNull(mBeanInfo);
-    Object flushCount = mBeanServer.getAttribute(oname, "Count");
-    Assert.assertTrue((long) flushCount >= 0);
-  }
-
-  private void validateVolumesList(String userName,
-      Set<String> expectedVolumes) throws Exception {
-
-    int expectedCount = 0;
-    Iterator<? extends OzoneVolume> volumeIterator =
-        objectStore.listVolumesByUser(userName, "", "");
-
-    while (volumeIterator.hasNext()) {
-      OzoneVolume next = volumeIterator.next();
-      Assert.assertTrue(expectedVolumes.contains(next.getName()));
-      expectedCount++;
-    }
-
-    Assert.assertEquals(expectedVolumes.size(),  expectedCount);
-  }
-
 
   /**
    * Stop the current leader OM.
